Implement solution of knights travail project
The graph is implicit. Vertices are represented by array of [x, y]
coordinates. A queue is used to explore different paths, the first loop
checks if one of the possible moves from the starting position found the
target. If not the algorithm repeats until a move lands on the target.
Since each loop explores all the possible moves from the current
position we can consider this algorithm a breadth-first-search, a level
order traversal. Each loop corresponds to a level. When a match is found
it means that the last move reached the target so all the paths that are
currently inserted into @queue lead to the target in the least amount of
moves, hence we retrieve one path with #shift.